P256k1PrivKey: add nested P256K1PrivKeyDefault implementation
* Migrate most of -ffm's PrivKeyPojo class into P256k1PrivKey.P256K1PrivKeyDefault
* Have PrivKeyPojo extend P256k1PrivKey.P256K1PrivKeyDefault (for MemorySegment stuff)
* Add P256k1PrivKey.of(BigInteger) convenience factory method
